Addressing comments
I had a tough time with this step. I made some more changes to try make the flow better on line 171-173, and rephrased around line 188. I think the thing that is critically missing in this last paragraph is how this all relates back to "Design with the end in mind". I'm tempted to think the filtering and searching point could be moved to rule 5 (Lines 185 - 189). At the end of rule 2 I can elaborate more on the practicality of persistent identifiers, how they are interconnected and help with 'finding' as a means for connecting back to the central theme of the rule. Let me know what you think of the idea and I can create a separate pull to try to communicate what I am thinking for this.