Update 10-findable.tex
I removed the brackets on line 200. I put that in because I was worried that not all readers would know what you mean about textual structures and it feels sort-of implied in the way it is currently written that the reader already knows. This document is the first time I'm seeing the term 'textual structure' to describe headings and text-organization based navigation mechanisms, but I know less about this topic - so if you feel like it is sufficiently defined, I am satisfied.

With regard to the changes on line 499 & 500. I rephrased my point - but I am willing to revert it to the original. I'll admit that the phrase "newcomers feel more welcome" was a bit triggering for me. One issue that I have with big, powerful institutions (like UCLA, but many others) is that when they think of diversity and inclusion of new people, they focus on making people feel welcome as if that is the most important issue with regard to individual retention. The community focuses on being extra polite and courteous and 'welcoming' to new people. This is fine, but at the same time institutions tend towards skepticism to new ideas. People in power wont always go out of their way to understand where the newcomers are coming from, they often expect it to be explained to them. For BIPOC entering technical fields trying to make transformative change: when they bring up ideas rooted in social justice or inclusion or anything that goes against the status quo, it is often viewed with a knee-jerk hostility. So the end result is a situation where a newcomer feels welcome in person, but their opinions and ideas are not valued in practice, something they consistently experience throughout their life. Anyway - this long-winded comment is based on context that might be only peripherally relevant to what we are writing here. That said, I always try to insert a point of caution when using the term "welcoming" in the context of newcomers in a field. I tried to reflect this in my revision. Let me know what you think - I am also happy to reverting it back to the original language if what I wrote is still unclear or out of context.